requirement.txt
asgiref==3.2.7
autopep8==1.5.2
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==3.0.6
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-registration==3.1
django-taggit==1.3.0
idna==2.9
phonenumbers==8.12.4
Pillow==7.1.2
pycodestyle==2.6.0
pytz==2020.1
requests==2.23.0
South==1.0.2
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
